requirements.txt
pysam>=0.8.2.1
biopython
pybedtools>=0.6.8
scipy>=0.15.1
numpy
nose>=1.3.0
progressbar2
pandas>=0.14.1
colorlog
mirtop
pyyaml
pytz
dateutils
